+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   breaking change if you have custom prefix without (registered) domain.
 - Add `createWithSender` to `JaegerTraceExporter` to allow use of `HttpSender`
   with extra configurations
+- Add an API `Functions.returnToString()`.
 
 ## 0.16.0 - 2018-09-14
 - Add APIs to register gRPC client and server views separately.